break the coredocument package
