--- coredocument/bootstrapper_test.go
@@ -1,14 +1,0 @@
-// +build unit
-
-package coredocument
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestBootstrapper_Bootstrap(t *testing.T) {
-	err := (&Bootstrapper{}).Bootstrap(map[string]interface{}{})
-	assert.Error(t, err, "Should throw an error because of empty context")
-}
