+++ documents/model.go
@@ -1,10 +1,9 @@
 package documents
 
 import (
+	"github.com/centrifuge/centrifuge-protobufs/gen/go/coredocument"
 	"github.com/centrifuge/go-centrifuge/storage"
 	"github.com/centrifuge/precise-proofs/proofs/proto"
-
-	"github.com/centrifuge/centrifuge-protobufs/gen/go/coredocument"
 )
 
 // Model is an interface to abstract away model specificness like invoice or purchaseOrder
