+++ documents/model_test.go
@@ -22,6 +22,7 @@ import (
 
 var ctx = map[string]interface{}{}
 var ConfigService config.Service
+var cfg config.Configuration
 
 func TestMain(m *testing.M) {
 	ibootstappers := []bootstrap.TestBootstrapper{
@@ -36,6 +37,7 @@ func TestMain(m *testing.M) {
 	ctx[identity.BootstrappedIDService] = &testingcommons.MockIDService{}
 	bootstrap.RunTestBootstrappers(ibootstappers, ctx)
 	ConfigService = ctx[config.BootstrappedConfigStorage].(config.Service)
+	cfg = ctx[bootstrap.BootstrappedConfig].(config.Configuration)
 	result := m.Run()
 	bootstrap.RunTestTeardown(ibootstappers)
 	os.Exit(result)
