+++ documents/validator.go
@@ -1,8 +1,17 @@
 package documents
 
 import (
+	"fmt"
+
+	"github.com/centrifuge/centrifuge-protobufs/gen/go/coredocument"
+	"github.com/centrifuge/go-centrifuge/anchors"
+	"github.com/centrifuge/go-centrifuge/centerrors"
+	"github.com/centrifuge/go-centrifuge/coredocument"
+	"github.com/centrifuge/go-centrifuge/crypto"
 	"github.com/centrifuge/go-centrifuge/errors"
+	"github.com/centrifuge/go-centrifuge/identity"
 	"github.com/centrifuge/go-centrifuge/utils"
+	"github.com/ethereum/go-ethereum/common/hexutil"
 )
 
 // Validator is an interface every Validator (atomic or group) should implement
@@ -38,4 +47,341 @@ type ValidatorFunc func(old, new Model) error
 // function and returns the results
 func (vf ValidatorFunc) Validate(old, new Model) error {
 	return vf(old, new)
+}
+
+// UpdateVersionValidator validates if the new core document is properly derived from old one
+func UpdateVersionValidator() Validator {
+	return ValidatorFunc(func(old, new Model) error {
+		if old == nil || new == nil {
+			return errors.New("need both the old and new model")
+		}
+
+		oldCD, err := old.PackCoreDocument()
+		if err != nil {
+			return errors.New("failed to fetch old core document: %v", err)
+		}
+
+		newCD, err := new.PackCoreDocument()
+		if err != nil {
+			return errors.New("failed to fetch new core document: %v", err)
+		}
+
+		checks := []struct {
+			name string
+			a, b []byte
+		}{
+			{
+				name: "cd_document_identifier",
+				a:    oldCD.DocumentIdentifier,
+				b:    newCD.DocumentIdentifier,
+			},
+
+			{
+				name: "cd_previous_version",
+				a:    oldCD.CurrentVersion,
+				b:    newCD.PreviousVersion,
+			},
+
+			{
+				name: "cd_current_version",
+				a:    oldCD.NextVersion,
+				b:    newCD.CurrentVersion,
+			},
+
+			{
+				name: "cd_previous_version",
+				a:    oldCD.DocumentRoot,
+				b:    newCD.PreviousRoot,
+			},
+		}
+
+		for _, c := range checks {
+			if !utils.CheckMultiple32BytesFilled(c.a, c.b) {
+				err = errors.AppendError(err, NewError(c.name, "missing identifiers"))
+				continue
+			}
+
+			if !utils.IsSameByteSlice(c.a, c.b) {
+				err = errors.AppendError(err, NewError(c.name, "mismatched"))
+			}
+		}
+
+		if utils.IsEmptyByteSlice(newCD.NextVersion) {
+			err = errors.AppendError(err, NewError("cd_next_version", centerrors.RequiredField))
+		}
+
+		return err
+	})
+}
+
+// getCoreDocument takes an model and returns the core document of the model
+func getCoreDocument(model Model) (*coredocumentpb.CoreDocument, error) {
+	if model == nil {
+		return nil, errors.New("nil model")
+	}
+
+	cd, err := model.PackCoreDocument()
+	if err != nil {
+		return nil, errors.New("failed to pack core document: %v", err)
+	}
+
+	return cd, nil
+}
+
+// baseValidator validates the core document basic fields like identifier, versions, and salts
+func baseValidator() Validator {
+	return ValidatorFunc(func(_, model Model) error {
+		cd, err := getCoreDocument(model)
+		if err != nil {
+			return err
+		}
+
+		if cd == nil {
+			return errors.New("nil document")
+		}
+
+		if utils.IsEmptyByteSlice(cd.DocumentIdentifier) {
+			err = errors.AppendError(err, NewError("cd_identifier", centerrors.RequiredField))
+		}
+
+		if utils.IsEmptyByteSlice(cd.CurrentVersion) {
+			err = errors.AppendError(err, NewError("cd_current_version", centerrors.RequiredField))
+		}
+
+		if utils.IsEmptyByteSlice(cd.NextVersion) {
+			err = errors.AppendError(err, NewError("cd_next_version", centerrors.RequiredField))
+		}
+
+		if utils.IsEmptyByteSlice(cd.DataRoot) {
+			err = errors.AppendError(err, NewError("cd_data_root", centerrors.RequiredField))
+		}
+
+		// double check the identifiers
+		isSameBytes := utils.IsSameByteSlice
+
+		// Problem (re-using an old identifier for NextVersion): CurrentVersion or DocumentIdentifier same as NextVersion
+		if isSameBytes(cd.NextVersion, cd.DocumentIdentifier) ||
+			isSameBytes(cd.NextVersion, cd.CurrentVersion) {
+			err = errors.AppendError(err, NewError("cd_overall", centerrors.IdentifierReUsed))
+		}
+
+		// lets not do verbose check like earlier since these will be
+		// generated by us mostly
+		salts := cd.CoredocumentSalts
+		if salts == nil ||
+			!utils.CheckMultiple32BytesFilled(
+				salts.CurrentVersion,
+				salts.NextVersion,
+				salts.DocumentIdentifier,
+				salts.PreviousRoot) {
+			err = errors.AppendError(err, NewError("cd_salts", centerrors.RequiredField))
+		}
+
+		return err
+	})
+}
+
+// signingRootValidator checks the existence of signing root
+// recalculates the signing root and compares with existing one
+func signingRootValidator() Validator {
+	return ValidatorFunc(func(_, model Model) error {
+		cd, err := getCoreDocument(model)
+		if err != nil {
+			return err
+		}
+
+		if utils.IsEmptyByteSlice(cd.SigningRoot) {
+			return errors.New("signing root missing")
+		}
+
+		tree, err := coredocument.GetDocumentSigningTree(cd)
+		if err != nil {
+			return errors.New("failed to calculate signing root: %v", err)
+		}
+
+		if !utils.IsSameByteSlice(cd.SigningRoot, tree.RootHash()) {
+			return errors.New("signing root mismatch")
+		}
+
+		return nil
+	})
+}
+
+// documentRootValidator checks the existence of document root
+// recalculates the document root and compares with existing one
+func documentRootValidator() Validator {
+	return ValidatorFunc(func(_, model Model) error {
+		cd, err := getCoreDocument(model)
+		if err != nil {
+			return err
+		}
+
+		if utils.IsEmptyByteSlice(cd.DocumentRoot) {
+			return errors.New("document root missing")
+		}
+
+		tree, err := coredocument.GetDocumentRootTree(cd)
+		if err != nil {
+			return errors.New("failed to calculate document root: %v", err)
+		}
+
+		if !utils.IsSameByteSlice(cd.DocumentRoot, tree.RootHash()) {
+			return errors.New("document root mismatch")
+		}
+
+		return nil
+	})
+}
+
+// readyForSignaturesValidator validates self signature
+// re-calculates the signature and compares with existing one
+// assumes signing_root is already generated and verified
+// Note: this needs to used only before document is sent for signatures from the collaborators
+func readyForSignaturesValidator(centIDBytes, priv, pub []byte) Validator {
+	return ValidatorFunc(func(_, model Model) error {
+		cd, err := getCoreDocument(model)
+		if err != nil {
+			return err
+		}
+
+		if len(cd.Signatures) != 1 {
+			return errors.New("expecting only one signature")
+		}
+
+		s := crypto.Sign(centIDBytes, priv, pub, cd.SigningRoot)
+		sh := cd.Signatures[0]
+		if !utils.IsSameByteSlice(s.EntityId, sh.EntityId) {
+			err = errors.AppendError(err, NewError("cd_entity_id", "entity ID mismatch"))
+		}
+
+		if !utils.IsSameByteSlice(s.PublicKey, sh.PublicKey) {
+			err = errors.AppendError(err, NewError("cd_public_key", "public key mismatch"))
+		}
+
+		if !utils.IsSameByteSlice(s.Signature, sh.Signature) {
+			err = errors.AppendError(err, NewError("cd_signature", "signature mismatch"))
+		}
+
+		return err
+	})
+}
+
+// signaturesValidator validates all the signatures in the core document
+// assumes signing root is verified
+// Note: can be used when during the signature request on collaborator side and post signature collection on sender side
+// Note: this will break the current flow where we proceed to anchor even signatures verification fails
+func signaturesValidator(idService identity.Service) Validator {
+	return ValidatorFunc(func(_, model Model) error {
+		cd, err := getCoreDocument(model)
+		if err != nil {
+			return err
+		}
+
+		if len(cd.Signatures) < 1 {
+			return errors.New("atleast one signature expected")
+		}
+
+		for _, sig := range cd.Signatures {
+			if erri := idService.ValidateSignature(sig, cd.SigningRoot); erri != nil {
+				err = errors.AppendError(
+					err,
+					NewError(
+						fmt.Sprintf("signature_%s", hexutil.Encode(sig.EntityId)),
+						fmt.Sprintf("signature verification failed: %v", erri)))
+			}
+		}
+
+		return err
+	})
+}
+
+// anchoredValidator checks if the document root matches the one on chain with specific anchorID
+// assumes document root is generated and verified
+func anchoredValidator(repo anchors.AnchorRepository) Validator {
+	return ValidatorFunc(func(_, new Model) error {
+		cd, err := getCoreDocument(new)
+		if err != nil {
+			return errors.New("failed to get core document: %v", err)
+		}
+
+		anchorID, err := anchors.ToAnchorID(cd.CurrentVersion)
+		if err != nil {
+			return errors.New("failed to get anchorID: %v", err)
+		}
+
+		docRoot, err := anchors.ToDocumentRoot(cd.DocumentRoot)
+		if err != nil {
+			return errors.New("failed to get document root: %v", err)
+		}
+
+		gotRoot, err := repo.GetDocumentRootOf(anchorID)
+		if err != nil {
+			return errors.New("failed to get document root from chain: %v", err)
+		}
+
+		if !utils.IsSameByteSlice(docRoot[:], gotRoot[:]) {
+			return errors.New("mismatched document roots")
+		}
+
+		return nil
+	})
+}
+
+// SignatureRequestValidator returns a validator group with following validators
+// base validator
+// signing root validator
+// signatures validator
+// should be used when node receives a document requesting for signature
+func SignatureRequestValidator(idService identity.Service) ValidatorGroup {
+	return PostSignatureRequestValidator(idService)
+}
+
+// PreAnchorValidator is a validator group with following validators
+// base validator
+// signing root validator
+// document root validator
+// signatures validator
+// should be called before pre anchoring
+func PreAnchorValidator(idService identity.Service) ValidatorGroup {
+	return ValidatorGroup{
+		PostSignatureRequestValidator(idService),
+		documentRootValidator(),
+	}
+}
+
+// PostAnchoredValidator is a validator group with following validators
+// PreAnchorValidator
+// anchoredValidator
+// should be called after anchoring the document/when received anchored document
+func PostAnchoredValidator(idService identity.Service, repo anchors.AnchorRepository) ValidatorGroup {
+	return ValidatorGroup{
+		PreAnchorValidator(idService),
+		anchoredValidator(repo),
+	}
+}
+
+// PreSignatureRequestValidator is a validator group with following validators
+// baseValidator
+// signingRootValidator
+// readyForSignaturesValidator
+// should be called after sender signing the document and before requesting the document
+func PreSignatureRequestValidator(centIDBytes, priv, pub []byte) ValidatorGroup {
+	return ValidatorGroup{
+		baseValidator(),
+		signingRootValidator(),
+		readyForSignaturesValidator(centIDBytes, priv, pub),
+	}
+}
+
+// PostSignatureRequestValidator is a validator group with following validators
+// baseValidator
+// signingRootValidator
+// signaturesValidator
+// should be called after the signature collection/before preparing for anchoring
+func PostSignatureRequestValidator(idService identity.Service) ValidatorGroup {
+	return ValidatorGroup{
+		baseValidator(),
+		signingRootValidator(),
+		signaturesValidator(idService),
+	}
 }
