+++ p2p/bootstrapper.go
@@ -5,7 +5,6 @@ import (
 	"github.com/centrifuge/go-centrifuge/config"
 	"github.com/centrifuge/go-centrifuge/config/configstore"
 	"github.com/centrifuge/go-centrifuge/documents"
-	"github.com/centrifuge/go-centrifuge/documents/genericdoc"
 	"github.com/centrifuge/go-centrifuge/errors"
 	"github.com/centrifuge/go-centrifuge/identity"
 	"github.com/centrifuge/go-centrifuge/p2p/receiver"
@@ -26,9 +25,9 @@ func (b Bootstrapper) Bootstrap(ctx map[string]interface{}) error {
 		return errors.New("configstore not initialised")
 	}
 
-	registry, ok := ctx[documents.BootstrappedRegistry].(*documents.ServiceRegistry)
+	docSrv, ok := ctx[documents.BootstrappedDocumentService].(documents.Service)
 	if !ok {
-		return errors.New("registry not initialised")
+		return errors.New("document service not initialised")
 	}
 
 	idService, ok := ctx[identity.BootstrappedIDService].(identity.Service)
@@ -36,15 +35,8 @@ func (b Bootstrapper) Bootstrap(ctx map[string]interface{}) error {
 		return errors.New("identity service not initialised")
 	}
 
-	genService, ok := ctx[genericdoc.BootstrappedGenService].(genericdoc.Service)
-	if !ok {
-		return errors.New("generic service is not initialised")
-	}
-
-	srv := &peer{config: cfgService, handlerCreator: func() *receiver.Handler {
-		return receiver.New(cfgService, registry, receiver.HandshakeValidator(cfg.GetNetworkID(), idService), genService)
+	ctx[bootstrap.BootstrappedPeer] = &peer{config: cfgService, handlerCreator: func() *receiver.Handler {
+		return receiver.New(cfgService, receiver.HandshakeValidator(cfg.GetNetworkID(), idService), docSrv)
 	}}
-	ctx[bootstrap.BootstrappedP2PServer] = srv
-	ctx[bootstrap.BootstrappedP2PClient] = srv
 	return nil
 }
