+++ p2p/bootstrapper_test.go
@@ -5,15 +5,13 @@ package p2p
 import (
 	"testing"
 
-	"github.com/centrifuge/go-centrifuge/documents/genericdoc"
-
 	"github.com/centrifuge/go-centrifuge/bootstrap"
 	"github.com/centrifuge/go-centrifuge/config"
 	"github.com/centrifuge/go-centrifuge/config/configstore"
 	"github.com/centrifuge/go-centrifuge/documents"
 	"github.com/centrifuge/go-centrifuge/identity"
 	"github.com/centrifuge/go-centrifuge/node"
-	testingcommons "github.com/centrifuge/go-centrifuge/testingutils/commons"
+	"github.com/centrifuge/go-centrifuge/testingutils/commons"
 	"github.com/centrifuge/go-centrifuge/testingutils/config"
 	"github.com/stretchr/testify/assert"
 )
@@ -31,16 +29,15 @@ func TestBootstrapper_Bootstrap(t *testing.T) {
 	m[bootstrap.BootstrappedConfig] = new(testingconfig.MockConfig)
 	m[config.BootstrappedConfigStorage] = cs
 	cs.On("GetConfig").Return(&configstore.NodeConfig{}, nil)
-	m[documents.BootstrappedRegistry] = documents.NewServiceRegistry()
 	ids := new(testingcommons.MockIDService)
 	m[identity.BootstrappedIDService] = ids
-	m[genericdoc.BootstrappedGenService] = genericdoc.DefaultService(nil, nil, nil)
+	m[documents.BootstrappedDocumentService] = documents.DefaultService(nil, nil, nil, documents.NewServiceRegistry())
 
 	err = b.Bootstrap(m)
 	assert.Nil(t, err)
 
-	assert.NotNil(t, m[bootstrap.BootstrappedP2PServer])
-	_, ok := m[bootstrap.BootstrappedP2PServer].(node.Server)
+	assert.NotNil(t, m[bootstrap.BootstrappedPeer])
+	_, ok := m[bootstrap.BootstrappedPeer].(node.Server)
 	assert.True(t, ok)
 
 	assert.NotNil(t, m[bootstrap.BootstrappedP2PClient])
