+++ p2p/receiver/handler.go
@@ -3,61 +3,38 @@ package receiver
 import (
 	"context"
 
-	"github.com/centrifuge/go-centrifuge/config"
-	"github.com/centrifuge/go-centrifuge/documents/genericdoc"
-	"github.com/centrifuge/go-centrifuge/identity"
-	"github.com/centrifuge/go-centrifuge/p2p/common"
-
-	"github.com/centrifuge/go-centrifuge/contextutil"
-
-	"github.com/golang/protobuf/proto"
-	"github.com/libp2p/go-libp2p-peer"
-	"github.com/libp2p/go-libp2p-protocol"
-
-	"github.com/centrifuge/centrifuge-protobufs/gen/go/coredocument"
 	"github.com/centrifuge/centrifuge-protobufs/gen/go/p2p"
 	"github.com/centrifuge/go-centrifuge/centerrors"
 	"github.com/centrifuge/go-centrifuge/code"
-	"github.com/centrifuge/go-centrifuge/coredocument"
+	"github.com/centrifuge/go-centrifuge/config"
+	"github.com/centrifuge/go-centrifuge/contextutil"
 	"github.com/centrifuge/go-centrifuge/documents"
 	"github.com/centrifuge/go-centrifuge/errors"
+	"github.com/centrifuge/go-centrifuge/identity"
+	"github.com/centrifuge/go-centrifuge/p2p/common"
 	pb "github.com/centrifuge/go-centrifuge/protobufs/gen/go/protocol"
+	"github.com/golang/protobuf/proto"
+	"github.com/libp2p/go-libp2p-peer"
+	"github.com/libp2p/go-libp2p-protocol"
 )
-
-// getService looks up the specific registry, derives service from core document
-func getServiceAndModel(registry *documents.ServiceRegistry, cd *coredocumentpb.CoreDocument) (documents.Service, documents.Model, error) {
-	if cd == nil {
-		return nil, nil, errors.New("nil core document")
-	}
-	docType, err := coredocument.GetTypeURL(cd)
-	if err != nil {
-		return nil, nil, errors.New("failed to get type of the document: %v", err)
-	}
-
-	srv, err := registry.LocateService(docType)
-	if err != nil {
-		return nil, nil, errors.New("failed to locate the service: %v", err)
-	}
-
-	model, err := srv.DeriveFromCoreDocument(cd)
-	if err != nil {
-		return nil, nil, errors.New("failed to derive model from core document: %v", err)
-	}
-
-	return srv, model, nil
-}
 
 // Handler implements protocol message handlers
 type Handler struct {
-	registry           *documents.ServiceRegistry
 	config             config.Service
 	handshakeValidator ValidatorGroup
-	genericService     genericdoc.Service
+	docSrv             documents.Service
 }
 
 // New returns an implementation of P2PServiceServer
-func New(config config.Service, registry *documents.ServiceRegistry, handshakeValidator ValidatorGroup, genericService genericdoc.Service) *Handler {
-	return &Handler{registry: registry, config: config, handshakeValidator: handshakeValidator, genericService: genericService}
+func New(
+	config config.Service,
+	handshakeValidator ValidatorGroup,
+	docSrv documents.Service) *Handler {
+	return &Handler{
+		config:             config,
+		handshakeValidator: handshakeValidator,
+		docSrv:             docSrv,
+	}
 }
 
 // HandleInterceptor acts as main entry point for all message types, routes the request to the correct handler
@@ -139,12 +116,12 @@ func (srv *Handler) HandleRequestDocumentSignature(ctx context.Context, peer pee
 // Existing signatures on the document will be verified
 // Document will be stored to the repository for state management
 func (srv *Handler) RequestDocumentSignature(ctx context.Context, sigReq *p2ppb.SignatureRequest) (*p2ppb.SignatureResponse, error) {
-	svc, model, err := getServiceAndModel(srv.registry, sigReq.Document)
+	model, err := srv.docSrv.DeriveFromCoreDocument(sigReq.Document)
 	if err != nil {
-		return nil, centerrors.New(code.DocumentInvalid, err.Error())
+		return nil, errors.New("failed to derive from core doc: %v", err)
 	}
 
-	signature, err := svc.RequestDocumentSignature(ctx, model)
+	signature, err := srv.docSrv.RequestDocumentSignature(ctx, model)
 	if err != nil {
 		return nil, centerrors.New(code.Unknown, err.Error())
 	}
@@ -180,12 +157,12 @@ func (srv *Handler) HandleSendAnchoredDocument(ctx context.Context, peer peer.ID
 
 // SendAnchoredDocument receives a new anchored document, validates and updates the document in DB
 func (srv *Handler) SendAnchoredDocument(ctx context.Context, docReq *p2ppb.AnchorDocumentRequest, senderID []byte) (*p2ppb.AnchorDocumentResponse, error) {
-	svc, model, err := getServiceAndModel(srv.registry, docReq.Document)
+	model, err := srv.docSrv.DeriveFromCoreDocument(docReq.Document)
 	if err != nil {
-		return nil, centerrors.New(code.DocumentInvalid, err.Error())
+		return nil, errors.New("failed to derive from core doc: %v", err)
 	}
 
-	err = svc.ReceiveAnchoredDocument(ctx, model, senderID)
+	err = srv.docSrv.ReceiveAnchoredDocument(ctx, model, senderID)
 	if err != nil {
 		return nil, centerrors.New(code.Unknown, err.Error())
 	}
@@ -221,7 +198,7 @@ func (srv *Handler) HandleGetDocument(ctx context.Context, peer peer.ID, protoc 
 
 // GetDocument receives document identifier and retrieves the corresponding CoreDocument from the repository
 func (srv *Handler) GetDocument(ctx context.Context, docReq *p2ppb.GetDocumentRequest) (*p2ppb.GetDocumentResponse, error) {
-	model, err := srv.genericService.GetCurrentVersion(ctx, docReq.DocumentIdentifier)
+	model, err := srv.docSrv.GetCurrentVersion(ctx, docReq.DocumentIdentifier)
 	if err != nil {
 		return nil, err
 	}
